Prepare UI Build Workflow
Add manual trigger for building the perfetto UI without releasing it

- Also package README.md to keep 'ui' file structure

Update build-perfetto-ui.yml to include generate proto defs

Daniel recently started using these so we need to add them to the Perfetto zip for Windows.

Update perfetto-ui.yml to add new protos dependency

Also made this so it can be invoked manually.  In this particular case, the Perfetto repo did not change, so this job was not triggered automatically.  So we could either push a noop change or add workflow_dispatch so we can trigger it manually in such cases.

Ensure consistent short sha in perfetto ui zip

Specify value of 14 when getting the short commit sha to use
in the perfetto ui zip. This ensures we get the same length
for ay given commit, regardless of which version of git
is being used. In 2016, git was updated to use a default
short sha length based on the size of the repo. That means
that --short can return a different length depending on the
git version. 14 should be more than plenty to ensure
uniqueness

Build for all commits to the branch (#146)

The ui filter was preventing non ui changes from triggering a
perfetto build. But we need a new build for any change, regardless
if the ui part of the build is changing or not. The issue was observed
with an sdk only change.