Fix PxSpan assertion errors
Do not take the new calculated width
when it is zero because the view is
temporarily not visible.

Don't update overview timeline panel when invisible (#51)

In the context of an embedding application, the Perfetto UI sometimes is
not visible and the client rect of the overview timeline panel can be
zero. In that case, creating the PxSpan fails its assertion. Fix that
by just not updating until eventually the client rect is available.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>