Parameterize the error handling
Extract the error handler that shows the
error dialog into the globals where other
points of customization are implemented,
while maintaining compatibility for contexts
such as the traceconv worker that doesn't
use the frontend globals.

Also add an API to force the error dialog
for the case when an external agent needs
to show it.