try: support multiple instantiation of Perfetto
Add a build step to copy the transpiled code
into into the package so that it can be
installed multiple times in an application
without sharing one copy via symlink
(and also because yarn breaks the
symlink when copying the package into
the node_modules/ directory).

Refactor the engine-ready callbacks into
the globals so that they may be replicated
along with the globals.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>