Added generated Perfetto SDK files
These are only added to the upstream Perfetto repo release branches.
They are generated by running running ./tools/gen_amalgamated.

These should be regenerated each time we take in changes from upstream.
I will add that to our readme in Sokatoa.

Update the perfetto sdk to v49 (#145)

Updating just the sdk, and not the full perfetto repository.
The update is needed for the device generation of events starting
on different threads. With the old perfetto we did loose some
of the events.
Note that the perfetto is designed to be compatible in respect
to the used sdk.