fix: area selection in side-by-side timelines
Perfetto UI does not account for the
possibility of the tracks container panel
not being aligned to the left edge of the
browser window, so it assumes that the
track shell width is sufficient to offset
the track location, but it isn't nearly
enough when there's a whole other
trace viewer to its left.

Update the calculation of covered
tracks to offset by the actual left
edge position of the panel container.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>