Add support for track descriptions
Add a description to the track data. Try to load it, where the schema
has one, from the database. Otherwise provide some interpretive text.
The latter descriptions are derived in a few different ways:

- database schema descriptions in the Perfetto documentation
- documentation from related systems (e.g., proc(5) manpage)
- mimic AGI (e.g., the display of the number of threads in a process)

Also fix the build script to account for the output directory that it
needs to clean now being a real directory and not a symbolic link.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>