Fix scrolling to tracks in nested groups
The introduction of nested groups means that, in addition to tracks, now
sometimes groups in the hierarchy will not be rendered yet when we want
to reveal a track. So add a similar mechanism in the frontend local
state to notify a group upon its rendering that it needs to be expanded
because the UI is trying to reveal a track within it.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>