Fix regression in adding thread state tracks
Before the introduction of "idle threads" groups in processes, the
TrackDecider::addThreadStateTracks() method was lenient in looking up
the group in which to create a thread state track, in case the thread
was suppressed on account of not having any scheduling data. The idle
groups feature broke that by asserting the existence of the track group.
This commit restores the previous behaviour under the new nomenclature.

Also rename a duplicate column name in the process groups query to avoid
confusion and possible bugs caused by the ambiguity.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>