Close the error dialog if it is opened before setting error handler
A custom error handler may be set by an embedding application via the
globals. However, an error may have already been trapped and presented
after Perfetto was loaded but before it was rendered. In this case, the
dialog was already created and ready to present when Perfetto was
rendered, after the embedding application set a different error handler.

Now when the error handler is set, any pending error dialog is closed or
prevented from being shown.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>