Ensure consistent short sha in perfetto ui zip
Specify value of 4 when getting the short commit sha to use
in the perfetto ui zip. This ensures we get the same length
for ay given commit, regardless of which version of git
is being used. In 2016, git was updated to use a default
short sha length based on the size of the repo. That means
that --short can return a different length depending on the
git version. By using --short=4 we are guranteed a length
that will not vary based on git version. Note that the actual
length returned can be longer than four and often will, due
to the need to guarantee it's unique for that given commit.

We are updating the consumer of the zip (sokatoa) to also
use --short=4