Render annotation flags over track group headers
All track rendering was done on a big scrolling canvas element that sits
behind the track dividing lines and shells. That is fine for rendering
stuff in the track content area of the scrolling tracks, but the track
content area of the headers of expanded groups does not scroll with the
rest. They're not even a part of the same static layout, instead being
positioned relatively and moreover sticking to the top of the panel as
tracks scroll up beneath them. Because they are in a different layout
to the canvas, they occlude anything painted on the canvas, so painting
in the expanded group headers has no effect.

Consequently, it is necessary to give each expanded group header its own
canvas on which we can draw content such as the stems of note flags.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>