Always open the details of a new selection (#129)
Ensure that whenever the selection changes and the new selection exists,
show it in the details panel even if the user had previously closed it.

Add a global flag to allow the host application to disable this
behaviour, for example on a user preference setting.

Includes incidental fixing of lint problems in edited files.

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>