Handle failures in initialization of filtered-out async slice tracks
The initial bounds-change handling of an async slice track performs
multiple queries asynchronously, in sequence, awaiting the results of
each in turn. During one of those waits, a host application embedding
Perfetto can filter that track out so that it is no longer in the tracks
table in global state and then accessing that state to formulate the
next SQL query throws an exception. Trap that exception and bail the
initialization of the track controller.

For android-graphics/soktatoa#2949

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>